Extract page view
